Make LispyNum be a defined-size type.
Without this, behaviour can be different on
different systems (even in practice: 32 vs 64 bit
systems).

+++ playground/lispy-hs/Lispy/Interpret.hs
@@ -51,7 +51,8 @@ pureBuiltinFunction TableFromPairs vals =
   case varargPairs vals of
     Nothing -> error "Odd number of arguments to table-from-pairs"
     Just pairs -> ImmTableValue (Map.fromList pairs)
-pureBuiltinFunction TableSize [ImmTableValue m] = NumberValue (Map.size m)
+pureBuiltinFunction TableSize [ImmTableValue m] =
+  NumberValue (fromIntegral (Map.size m))
 pureBuiltinFunction TableViewKey [ImmTableValue m, k] =
   ImmTableViewValue (createMapIterator k m)
 pureBuiltinFunction TableUnView [ImmTableViewValue i] =
